Updated store factories in samples
