+++ sample/coroutines/shared/src/commonMain/kotlin/com/arkivanov/mvikotlin/sample/coroutines/shared/main/store/add/AddStore.kt
@@ -1,18 +1,18 @@
-package com.arkivanov.mvikotlin.sample.reaktive.shared.main.store
+package com.arkivanov.mvikotlin.sample.coroutines.shared.main.store.add
 
 import com.arkivanov.mvikotlin.core.store.Store
 import com.arkivanov.mvikotlin.core.utils.JvmSerializable
+import com.arkivanov.mvikotlin.sample.coroutines.shared.main.store.add.AddStore.Intent
+import com.arkivanov.mvikotlin.sample.coroutines.shared.main.store.add.AddStore.Label
+import com.arkivanov.mvikotlin.sample.coroutines.shared.main.store.add.AddStore.State
 import com.arkivanov.mvikotlin.sample.database.TodoItem
-import com.arkivanov.mvikotlin.sample.reaktive.shared.main.store.AddStore.Intent
-import com.arkivanov.mvikotlin.sample.reaktive.shared.main.store.AddStore.Label
-import com.arkivanov.mvikotlin.sample.reaktive.shared.main.store.AddStore.State
 
 internal interface AddStore : Store<Intent, State, Label> {
 
     // Serializable only for exporting events in Time Travel, no need otherwise.
     sealed class Intent : JvmSerializable {
         data class SetText(val text: String) : Intent()
-        object Add : Intent()
+        data object Add : Intent()
     }
 
     data class State(
