+++ sample/coroutines/shared/src/commonTest/kotlin/com/arkivanov/mvikotlin/sample/coroutines/shared/details/store/DetailsStoreTest.kt
@@ -140,12 +140,11 @@ class DetailsStoreTest {
 
     private fun createStore() {
         store =
-            DetailsStoreFactory(
-                storeFactory = DefaultStoreFactory(),
+            DefaultStoreFactory().detailsStore(
                 database = database,
                 mainContext = Dispatchers.Unconfined,
                 ioContext = Dispatchers.Unconfined,
                 itemId = itemId,
-            ).create()
+            )
     }
 }
